Arreglo error de inicializacion de tests

+++ test/test.js
@@ -36,10 +36,12 @@ describe('exec-to-html', function(){
         var dest = Path.normalize(dirbase+'/pro1/');
         var destNM = dest+'/node_modules';
         var bakNM = dirtemp+'/node_modules';
+        var existsNM=false;
         Promises.start(function(){
             return fs.exists(destNM);
-        }).then(function(existsNM) {
-            if(existsNM) { /*console.log(destNM, "->", bakNM);*/ return fs.rename(destNM, bakNM); }
+        }).then(function(exists) {
+            existsNM = exists;
+            if(existsNM) { return fs.rename(destNM, bakNM); }
         }).then(function() {
             return fs.remove(dirbase);
         }).then(function(){
@@ -47,8 +49,7 @@ describe('exec-to-html', function(){
         }).then(function(){
             return fs.rename(dest+'dot-git', dest+'.git');
         }).then(function(){
-            //console.log(bakNM, "->", destNM);
-            return fs.rename(bakNM, destNM);
+            if(existsNM) { return fs.rename(bakNM, destNM); }
         }).then(function() {
             done();
         }).catch(function(err){
